tgbot update and new user funcs for plantchecker

+++ plantchecker.py
@@ -20,6 +20,18 @@ connection = connect(
 
 class Plantchecker():
 
+    def get_user_id(login:str = None):
+        if login is None:
+            return "Мы вас не узнали. Укажите ваш логин"
+        with connection.cursor() as cursor:
+            cursor.execute('select id from users where login = "{login}"'.format(login=login))
+            user_id = cursor.fetchall()
+        cursor.close()
+        if len(user_id) == 0:
+            return 'Error, no id'
+        else:
+            return user_id[0][0]
+
     def check_user_plants(login: str = None, user_id: int = None):
         global connection
         if user_id is None and login is None:
@@ -37,22 +49,23 @@ class Plantchecker():
             cursor.execute('select plantname from userplants where user="{id}"'.format(id=user_id))
             userplants_list = [('').join(x) for x in cursor.fetchall()]
             if len(userplants_list) == 0:
-                return ('У вас еще нет растений')
+                return ("У вас еще нет растений")
 
         user_plants = {}
         for name in userplants_list:
             plant_card = {}
             for x in columnlist:
-                with connection.cursor() as cursor:
-                    cursor.execute(
-                        'select {column} from userplants where user={id} and plantname="{name}"'.format(column=x,
-                                                                                                        id=user_id,
-                                                                                                        name=name))
-                    data = cursor.fetchall()[0][0]
-                    plant_card[x] = data
+                columns = ['id','user','plantname','plantspec']
+                if x in columns:
+                    with connection.cursor() as cursor:
+                        cursor.execute('select {column} from userplants where user={id} and plantname="{name}"'.format(column=x, id=user_id, name=name))
+                        data = cursor.fetchall()[0][0]
+                        plant_card[x] = data
 
             user_plants.setdefault(name, plant_card)
         cursor.close()
+        user_plants = json.dumps(user_plants, indent=4, ensure_ascii=False)
+        print(type(user_plants))
         return user_plants
 
     def user_plantcard(plant_id: int = None, plantname: str = ''):
@@ -83,12 +96,13 @@ class Plantchecker():
         if plantspec not in plant_dict:
             return 'Не опознали вид растения'
 
-        if type(last_watering) == str and (len(last_watering) == 10) and last_watering.find('.') == 2:
-            plant_data["last_watering"] = datetime.datetime.strptime(last_watering, "%d.%m.%Y").date()
+        if type(last_watering) == str and (len(last_watering) == 10) and last_watering.find('-') == 2:
+            plant_data["last_watering"] = last_watering
         elif last_watering is None:
-            plant_data['last_watering'] = datetime.date.today()
+            date = datetime.date.today()
+            plant_data['last_watering'] = date.strftime("%d-%m-%Y")
         else:
-            return 'Неправильный формат даты. Используйте "dd.mm.yyyy".\n'
+            return 'Неправильный формат даты. Используйте "dd-mm-yyyy".\n'
 
         plant_data["last_fertiling"] = plant_data["last_watering"]
         plant_data['plantspec'] = plantspec
@@ -132,12 +146,13 @@ class Plantchecker():
     def add_plant_water(plantdata: json):
         global connection
         add_data = {}
+        plantdata = json.load(plantdata)
         if 'date' in plantdata:
             date = plantdata['date']
-            if type(date) == str and len(date) == 10 and date.find('.') == 2:
-                date = datetime.datetime.strptime(date, "%d.%m.%Y").date()
+            if type(date) == str and len(date) == 10 and date.find('-') == 2:
+                date = datetime.datetime.strptime(date, "%d-%m-%Y").date()
             else:
-                return 'Неправильный формат даты. Используйте строку "dd.mm.yyyy".\n'
+                return 'Неправильный формат даты. Используйте строку "dd-mm-yyyy".\n'
         else:
             date = datetime.date.today()
 
@@ -153,14 +168,15 @@ class Plantchecker():
                          where id = "{id}" or plantname = "{name}"'''.format(id=plant_id, name=plant))
             add_data['id'] = plant_id
             add_data['name'] = plant
-            add_data['date'] = date
+            add_data['date'] = date.strftime("%d-%m-%Y")
             add_string = '''update userplants set last_watering = %(date)s, 
                                     next_water= %(next_date)s where id = %(id)s or plantname = %(name)s'''
 
             plantsql = cursor.fetchall()
             this_month = date.month
             gap = plantsql[0][1] if this_month in range(3, 10) else plantsql[0][2]
-            add_data['next_date'] = date + datetime.timedelta(days=gap)
+            next_date = date + datetime.timedelta(days=gap)
+            add_data['next_date'] = next_date.strftime("%d-%m-%Y")
 
             try:
                 cursor.execute(add_string, add_data)
@@ -176,13 +192,14 @@ class Plantchecker():
     def add_plant_fertile(plantdata: json):
         global connection
         add_data = {}
+        plantdata = json.load(plantdata)
         if 'date' in plantdata:
             date = plantdata['date']
-            if type(date) == str and len(date) == 10 and date.find('.') == 2:
-                date = datetime.datetime.strptime(date, "%d.%m.%Y").date()
+            if type(date) == str and len(date) == 10 and date.find('-') == 2:
+                date = datetime.datetime.strptime(date, "%d-%m-%Y").date()
 
             else:
-                return 'Неправильный формат даты. Используйте строку "dd.mm.yyyy".\n'
+                return 'Неправильный формат даты. Используйте строку "dd-mm-yyyy".\n'
         else:
             date = datetime.date.today()
 
@@ -198,14 +215,15 @@ class Plantchecker():
                          where id = "{id}" or plantname = "{name}"'''.format(id=plant_id, name=plant))
             add_data['id'] = plant_id
             add_data['name'] = plant
-            add_data['date'] = date
+            add_data['date'] = date.strftime("%d-%m-%Y")
             add_string = '''update userplants set last_fertiling = %(date)s, 
                                     next_fertile= %(next_date)s where id = %(id)s or plantname = %(name)s'''
 
             plantsql = cursor.fetchall()
             this_month = date.month
             gap = plantsql[0][1] if this_month in range(3, 10) else plantsql[0][2]
-            add_data['next_date'] = date + datetime.timedelta(days=gap)
+            next_date = date + datetime.timedelta(days=gap)
+            add_data['next_date'] = next_date.strftime("%d-%m-%Y")
 
             try:
                 cursor.execute(add_string, add_data)
@@ -221,84 +239,64 @@ class Plantchecker():
 
     def delete_plant(login: str = None, plantname: str = None, **kwargs):
         global connection
-        plant_id = kwargs.get("plant_id", default=None)
-        user_id = kwargs.get("user_id", default=None)
-
-        if user_id is None and login is None:
+        if login is None:
             return "Мы вас не узнали. Укажите ваш логин"
-        elif user_id is None:
-            with connection.cursor() as cursor:
-                cursor.execute('select id from users where login = "{login}"'.format(login=login))
-                user_id = cursor.fetchone()
-                user_id = user_id[0]
-            cursor.close()
+        user_id = Plantchecker.get_user_id(login)
+        if type(user_id) == str:
+            user_id = 0
+
+        plant_id = kwargs.get("plant_id", default=None) if len(kwargs) !=0 else 0
 
         with connection.cursor() as cursor:
-            cursor.execute('select plantname from userplants where user="{id}"'.format(id=user_id))
+            cursor.execute('select plantname from userplants where user="{id}" and (plantname="{plantname}" or id={plant_id})'.format(plantname=plantname, id=user_id, plant_id=plant_id))
             plantlist = [''.join(x) for x in cursor.fetchall()]
+
             if len(plantlist) == 0:
-                return 'У вас еще нет растений'
-            cursor.execute('delete from userplants where user="{id}" and (plantname="{plantname}" '
-                           'or plant_id="{plant_id}"'.format(plantname=plantname, id=user_id, plant_id=plant_id))
+                return "Растение не существует"
+
+            cursor.execute('delete from userplants where user="{id}" and (plantname="{plantname}" or id="{plant_id}")'.format(plantname=plantname, id=user_id, plant_id=plant_id))
             connection.commit()
-            cursor.execute('select plantname from userplants where user="{id}" and '
-                           '(plantname="{plantname}"'.format(plantname=plantname, id=user_id))
+            cursor.execute('select plantname from userplants where user="{id}" and (plantname="{plantname}" or id={plant_id})'.format(plantname=plantname, id=user_id, plant_id=plant_id))
 
             plantlist = [''.join(x) for x in cursor.fetchall()]
-            cursor.close()
+        cursor.close()
         return 'Растение удалено' if len(plantlist) == 0 else 'Что-то случилось, повторите запрос'
 
     def add_user(login: str = None, name: str = None, **kwargs):
         global connection
         if name is None and login is None:
-            return 100
+            return "100 - not enough parameters"
         with connection.cursor() as cursor:
             cursor.execute('select * from users where login="{login}"'.format(login=login))
             users = [''.join(str(x)) for x in cursor.fetchall()]
             if len(users) == 0:
                 cursor.execute('insert into users values (Null,"{login}", "{name}")'.format(login=login, name=name))
                 connection.commit()
-                return 200
+                return "200 - user added"
 
             elif len(users) == 1:
-                return 201
+                return "201 - user is already added"
 
             else:
-                return 400
+                return "400 - error"
 
         cursor.close()
-
-    def check_user(login: str = None, name: str = None):
-        global connection
-        if name is None and login is None:
-            return 100
-        with connection.cursor() as cursor:
-            cursor.execute('select * from users where login="{login}"'.format(login=login))
-            users = [''.join(str(x)) for x in cursor.fetchall()]
-
-            if len(users) == 1:
-                if name is not None:
-                    if users[2] != name or users[2] is None:
-                        cursor.execute('update users set name = "{name}" where login = "{login}"'.format(name = name, login = login))
-                        connection.commit()
-                        cursor.close()
-                        return 201
-
-                cursor.close()
-                return 200
-            elif len(users)==0:
-                return Plantchecker.add_user(login,name)
-            else:
-                return 400
 
     def delete_user(login: str):
         pass
 
 
 if __name__ == '__main__':
-    # with connection.cursor() as cursor:
-    #     cursor.execute('update users set name="anna" where login="linlynx";')
-    #     connection.commit()
-    # cursor.close()
+    pass
 
-    print(Plantchecker.check_user(login='56789'))
+    # path = os.path.abspath('test.json')
+    # with open(path) as file:
+    #     print(Plantchecker.add_plant_fertile(file))
+
+    # column = ['last_watering', "last_fertiling", "next_water", "next_fertile"]
+    # with connection.cursor() as cursor:
+    #     cursor.execute('select {column} from userplants where id=6'.format(column=column[0]))
+    #     datelist = cursor.fetchall()
+    #     print(type((datelist[0][0])))
+    #     # connection.commit()
+    # cursor.close()
